Aligning packed PWC messages.

+++ src/libphoton/photon_request.h
@@ -43,7 +43,7 @@
 #define REQUEST_FLAG_2PWC    (1<<8)
 
 #define MARK_DONE(e,s)         (sync_fadd(&e->tail, s, SYNC_RELAXED))
-#define EB_MSG_SIZE(s)         (sizeof(struct photon_eb_hdr_t) + s + sizeof(uintmax_t))
+#define EB_MSG_SIZE(s)         (sizeof(struct photon_eb_hdr_t) + s + sizeof(uint8_t))
 #define PROC_REQUEST_ID(p, id) (((uint64_t)p<<32) | id)
 #define IS_VALID_PROC(p)       ((p >= 0) && (p < _photon_nproc))
